requirements.txt
flask==3.1.0
requests==2.32.3
beautifulsoup4==4.13.3
google-genai>=1.0.0
google-api-python-client>=2.0.0
